+++ src/pages/Welcome/Welcome.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import Welcome1 from '../../features/welcome/Welcome1';
 import Welcome3 from '../../features/welcome/Welcome3';
+import Welcome4 from '../../features/welcome/Welcome4';
 import Welcome5 from '../../features/welcome/Welcome5';
 import Welcome6 from '../../features/welcome/Welcome6';
-import Welcome7 from '../../features/welcome/Welcome7';
 
 function Welcome() {
   return (
@@ -12,11 +12,11 @@ function Welcome() {
       {/* ////////////////////////////////////////////////////////////////////////////////////// */}
       <Welcome3 />
       {/* ////////////////////////////////////////////////////////////////////////////////////// */}
+      <Welcome4 />
+      {/* ////////////////////////////////////////////////////////////////////////////////////// */}
       <Welcome5 />
       {/* ////////////////////////////////////////////////////////////////////////////////////// */}
       <Welcome6 />
-      {/* ////////////////////////////////////////////////////////////////////////////////////// */}
-      <Welcome7 />
     </>
   );
 }
