+++ src/redux/modules/loginSlice.js
@@ -1,4 +1,4 @@
-import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
+import { createSlice } from "@reduxjs/toolkit";
 import { cookies } from "../../shared/cookies";
 
 const initialState = {
